dev (edit content): address feedback
