+++ core-web/libs/edit-content/src/lib/fields/dot-edit-content-text-area/dot-edit-content-text-area.component.html
@@ -1,7 +1,7 @@
 <textarea
     [id]="field.variable"
     [formControlName]="field.variable"
-    [attr.data-testId]="'text-area-' + field.variable"
+    [attr.data-testId]="field.variable"
     [style]="{
         'min-height': '18.75rem',
         resize: 'vertical'
