--- core-web/libs/edit-content/src/lib/fields/models/index.ts
@@ -1,13 +1,0 @@
-// Input type that you can select when creating the field
-export enum INPUT_TYPE {
-    TEXT = 'TEXT',
-    INTEGER = 'INTEGER',
-    FLOAT = 'FLOAT'
-}
-
-// This is to hold the options for the input type
-export interface InputTextOptions {
-    type: string;
-    inputMode: string;
-    step?: number;
-}
